+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Don't drop `Announce(Dislike)` activities.
+
 ## [3.2.0] - 2024-08-18
 
 ### Added